+++ dev_deps.ts
@@ -1,7 +1,0 @@
-export { assertEquals } from "https://deno.land/std@0.157.0/testing/asserts.ts";
-export {
-  assertSpyCall,
-  assertSpyCalls,
-  spy,
-} from "https://deno.land/std@0.157.0/testing/mock.ts";
-export { describe, it } from "https://deno.land/std@0.157.0/testing/bdd.ts";